Remove `@Async` from RequestScoped bean This might have caused a memory leak.
